requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
pymysql==1.1.0
cryptography==41.0.4
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
flask-caching==2.1.0
openai==1.3.5
pytest==7.4.3
lxml